Force type coercion for checkeq() template
On FreeBSD-11.0p7/amd64 with default compiler /usr/bin/CC (which is clang version 3.8.0)
ep-engine fails to compile because of template instantiation error

Below what cmake is using for /usr/bin/CC

  CXX_FLAGS =   -std=c++11 -g -Qunused-arguments -Wall -pedantic -Wredundant-decls \
                -fno-strict-aliasing -fvisibility=hidden -pthread -Werror=switch \
                -ftemplate-depth=900  -O3 -DNDEBUG -g -fPIC

  CXX_DEFINES = -DENTERPRISE_EDITION=1 -DHAVE_JEMALLOC -DPHOSPHOR_DISABLED=1 \
                -D_GNU_SOURCE=1 -D_POSIX_PTHREAD_SEMANTICS -D__EXTENSIONS__=1 \
                -D__STDC_FORMAT_MACROS -Dep_testsuite_xdcr_EXPORTS

Reviewed-on: http://review.couchbase.org/73188
Tested-by: Build Bot <build@couchbase.com>

+++ tests/ep_testsuite_xdcr.cc
@@ -1808,12 +1808,12 @@ static enum test_result test_setting_drift_threshold(ENGINE_HANDLE *h, ENGINE_HA
                      std::get<1>(conf).c_str(), data.first.data()),
                 "Expected set_param success");
 
-            checkeq(data.second.count(),
+            checkeq(int64_t(data.second.count()),
                     int64_t(get_ull_stat(h, h1, std::get<0>(conf).c_str(), nullptr)),
                     "Expected the stat to change to the new value");
 
             // The VB stat values are in nanoseconds
-            checkeq(std::chrono::nanoseconds(data.second).count(),
+            checkeq(int64_t(std::chrono::nanoseconds(data.second).count()),
                     int64_t(get_ull_stat(h, h1, std::get<2>(conf).c_str(), "vbucket-details 0")),
                     "Expected the VB stats to change to the new value");
         }
